+++ scripts/BruteForce/player.lua
@@ -5,23 +5,23 @@ local core = require("openmw.core")
 local I = require("openmw.interfaces")
 
 require("scripts.BruteForce.utils.consts")
-local L = require("scripts.BruteForce.bf_logic")
-local omw_utils = require("scripts.BruteForce.utils.openmw_utils")
+require("scripts.BruteForce.logic.onHit")
+require("scripts.BruteForce.logic.onUnlock")
 
 local sectionOnHit = storage.globalSection("SettingsBruteForce_onHit")
 local sectionOnUnlock = storage.globalSection("SettingsBruteForce_onUnlock")
 
 local function onObjectHit(o, var, res)
-    if not L.registerAttack(o) then return end
+    if not RegisterAttack(o) then return end
 
-    if L.attackMissed(o) or L.weaponTooWorn(o) then
+    if AttackMissed(o, self) or WeaponTooWorn(o, self) then
         if sectionOnHit:get("damageOnH2hMisses") then
-            L.damageIfH2h()
+            DamageIfH2h(self)
         end
         return
     end
 
-    L.damageIfH2h()
+    DamageIfH2h(self)
     core.sendGlobalEvent("checkJammedLock", { o = o, sender = self })
     -- check jammed lock in global script
     -- if it's OK, it will fire a tryUnlocking event back here
@@ -30,27 +30,27 @@ end
 local function lockWasntJammed(data)
     local o = data.o
 
-    if not L.unlock(o) then
+    if not Unlock(o, self) then
         -- lock got bent
         if sectionOnUnlock:get("enableWeaponWearAgainstBentLocks") then
-            L.wearWeapon(o, self)
+            WearWeapon(o, self)
         end
         return
     end
 
-    L.giveCurrWeaponXp()
-    L.wearWeapon(o, self)
+    GiveCurrWeaponXp(self)
+    WearWeapon(o, self)
 
     if sectionOnUnlock:get("triggerTraps") then
-        L.triggerTrap(o, self)
+        TriggerTrap(o, self)
     end
 
-    if omw_utils.objectIsOwned(o, self) then
-        L.alertNpcs()
+    if ObjectIsOwned(o, self) then
+        AlertNpcs(self)
     end
 
     if types.Container.objectIsInstance(o) then
-        L.damageContainerEquipment(o)
+        DamageContainerEquipment(o)
     end
 end
 
@@ -58,15 +58,15 @@ local function lockWasJammed(data)
     local o = data.o
 
     if sectionOnUnlock:get("enableWeaponWearAgainstBentLocks") then
-        L.wearWeapon(o, self)
+        WearWeapon(o, self)
     end
 
     if sectionOnUnlock:get("triggerTraps") then
-        L.triggerTrap(o, self)
+        TriggerTrap(o, self)
     end
 end
 
-omw_utils.checkDependencies(self, Dependencies)
+CheckDependencies(self, Dependencies)
 I.impactEffects.addHitObjectHandler(onObjectHit)
 
 return {