+++ scripts/BruteForce/utils/detection.lua
@@ -1,13 +1,11 @@
 -- Code taken from "SHOP - Store & House Owner Patrol (NPC in interiors AI overhaul) for OpenMW" by Łukasz "skrow42" Walczak
 
-local util      = require('openmw.util')
-
-local detection = {}
+local util = require('openmw.util')
 
 -- Offset vectors for LOS
-local vEye      = util.vector3(0, 0, 90)
-local vChest    = util.vector3(0, 0, 60)
-local vFeet     = util.vector3(0, 0, 15)
+local vEye   = util.vector3(0, 0, 90)
+local vChest = util.vector3(0, 0, 60)
+local vFeet  = util.vector3(0, 0, 15)
 
 ----------------------------------------------------------------------
 -- Line of Sight
@@ -24,7 +22,7 @@ local function clearRay(from, to, ignoreNpc, nearby, self)
     return not result.hit
 end
 
-function detection.canNpcSeePlayer(npc, self, nearby, maxDistance)
+function CanNpcSeePlayer(npc, self, nearby, maxDistance)
     if not self.cell then return false end
     if not (npc and npc:isValid()) then return false end
     if npc.cell ~= self.cell then return false end
@@ -51,7 +49,7 @@ end
 -- General distance
 ----------------------------------------------------------------------
 
-function detection.isWithinDistance(npc, self, maxDistance)
+function IsWithinDistance(npc, self, maxDistance)
     if not self.cell then return false end
     if not (npc and npc:isValid()) then return false end
     if npc.cell ~= self.cell then return false end
@@ -62,5 +60,3 @@ function detection.isWithinDistance(npc, self, maxDistance)
 
     return distance <= maxDistance
 end
-
-return detection