+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - New SFX for breaking and bending locks
 - Breaking trapped locks trigger traps
 - Improved ownership checking
+- Restructured code. Hopefully, everything has been left in tact
 
 ## 1.2.4